requirements.txt
numpy~=2.3.4
ursina~=7.0.0
pillow~=12.0.0
